+++ CHANGELOG.md
@@ -3,4 +3,5 @@
 ## 2016-01-12
 
 ### Added
+- Added main source file and converted the functions to TypeScript.
 - Initial commit.